+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `fillchars`, `foldtext`, and `smoothscroll` configurations.
 - Relaxed yaml dump quoting condition so that dates are not quoted
 - Update Stylua version from 0.15.1 → 2.1.0
+- Use `vim.deprecate` to show deprecate warnings
 
 ### Fixed